+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-Update to 1.21.4.
+Added Vibranium Items:
+
++ A better Captain America Shield 3d model !
+
+Add a lot more coming soon !